Convert to floating point

+++ src/models.rs
@@ -13,12 +13,12 @@ pub struct Account {
     pub id: i32,
     pub acct_name: String,
     pub acct_type: String,
-    pub debit_credit: i32,
+    pub debit_credit: f64,
 }
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct TransactionSplit {
-    pub amount: i32,
+    pub amount: f64,
     pub account: String,
     pub account_type: Option<String>,
 }
@@ -26,7 +26,7 @@ pub struct TransactionSplit {
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Transaction {
     pub date: String,
-    pub debit_credit: i32,
+    pub debit_credit: f64,
     pub acct_name: String,
     pub acct_type: String,
     pub acct_offset_name: String,
